requirements.txt
# FastAPI y dependencias
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
email-validator==1.1.3

# Base de datos
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
alembic==1.7.1

# Seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5

# Manejo de fechas y zonas horarias
pytz==2021.1

# Herramientas de desarrollo
black==21.7b0
isort==5.9.3
flake8==3.9.2
mypy==0.910

# Testing
pytest==6.2.5
pytest-cov==2.12.1

# Otras utilidades
requests==2.26.0
python-dotenv==0.19.0
apscheduler==3.7.0

# Documentación
mkdocs==1.2.2
mkdocs-material==7.2.6